shift things around abti
