+++ src/attribute_set/ffi.rs
@@ -1,0 +1,5 @@
+use super::AttributeSet;
+
+extern "C" {
+    pub fn AttributeSet_Free(set: *mut AttributeSet);
+}
