+++ src/attribute_set/mod.rs
@@ -1,0 +1,5 @@
+use std::ffi::c_void;
+
+pub mod ffi;
+
+pub type AttributeSet = *mut c_void;
