--- src/binding/mod.rs
@@ -1,6 +1,0 @@
-/*
- * Copyright FalkorDB Ltd. 2023 - present
- * Licensed under the Server Side Public License v1 (SSPLv1).
- */
-
-pub mod graph;
