+++ src/commands/ffi.rs
@@ -1,0 +1,1 @@
+extern "C" {}
