+++ src/schema/mod.rs
@@ -1,0 +1,6 @@
+#[repr(C)]
+#[derive(Copy, Clone, PartialEq)]
+pub enum SchemaType {
+    Node,
+    Edge,
+}
