+++ src/types/mod.rs
@@ -1,0 +1,6 @@
+pub type NodeID = i64;
+pub type EntityID = i64;
+pub type LabelID = i32;
+pub type SchemaID = i32;
+pub type RelationID = i32;
+pub type AttributeID = i32;
